fix(devtools): module getters not initialized
